refactor: 💡 변경 Prize 내에 검색기능 구성 및 PrizeSearchTool 제거
PrizeSearchTool이 Prize와 너무 깊게 맞물린 것 같아서 Prize 내로 해당 기능을 옮김
