+++ src/main/java/constant/Prize.java
@@ -3,6 +3,8 @@ package constant;
 import lotto.PrizeSearchTool;
 
 import java.text.NumberFormat;
+import java.util.HashMap;
+import java.util.Map;
 
 public enum Prize {
     NONE(0, false, 0),
@@ -21,12 +23,14 @@ public enum Prize {
 
     private final String description;
 
+    static final Map<Integer, Map<Boolean, Prize>> matchPrize = new HashMap<>();
+
     Prize(int match, boolean bonus, int money) {
         this.match = match;
         this.bonus = bonus;
         this.money = money;
         this.description = writeDescription();
-        PrizeSearchTool.updateSearchTool(this);
+        updateSearchTool(this);
     }
 
     private String writeDescription() {
@@ -39,6 +43,29 @@ public enum Prize {
         }
 
         return "";
+    }
+
+    private static void updateSearchTool(Prize prize) {
+        int match = prize.getMatch();
+        Map<Boolean, Prize> booleanPrize = matchPrize.getOrDefault(match, new HashMap<>());
+        booleanPrize.put(prize.isBonus(), prize);
+        matchPrize.put(match, booleanPrize);
+    }
+
+    public static Prize search(int match, boolean bonus) {
+        Map<Boolean, Prize> booleanPrize = matchPrize.get(match);
+
+        if (booleanPrize == null || booleanPrize.size() == 0) {
+            return Prize.NONE;
+        }
+
+        Prize prize = booleanPrize.get(bonus);
+
+        if (prize == null) {
+            return Prize.NONE;
+        }
+
+        return prize;
     }
 
     public int getMoney() {
