+++ src/main/java/lotto/Statistics.java
@@ -37,7 +37,7 @@ public class Statistics {
         Set<Integer> numbers = new HashSet<>(lotto.getNumbers());
         int match = getMatch(winNumbers, numbers);
         boolean bonus = isBonus(bonusNumber, numbers, match);
-        Prize prize = PrizeSearchTool.search(match, bonus);
+        Prize prize = Prize.search(match, bonus);
         pileUp(output, prize);
     }
